requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
gunicorn
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
psycopg-binary==3.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
werkzeug==3.0.4
zipp==3.20.2
